+++ NEWS.md
@@ -5,6 +5,7 @@
  * fix deparse in qae() and qe().
  * Deprecate ateval(), seval(), beval(), and "stringsubs".
  * add qs().
+ * add dereference and class-supplied function to pipes.
 
 # wrapr 1.0.2 2017-12-13